Just an archive of the project
Archive of the ada.rose-hulman.edu gitlab project.
This project was sponsored by Rolls Royce as our Senior Project. I worked on the CHT and CHT2 files independantly for the most part as most of my job in the group was to research vulnerabilities. CHT was build on socketCAN while CHT2 was built on from pyvit. CHT was dropped a few weeks in for a preference towards the ease of use of pyvit. The only non-working part of CHT2/hack is the UDS and IsoTP portions because of a time constraint. The grapher.py file is also old, the newest revision was lost but as this is an archive I will not be redoing the file.